refactor: move `File` and `Dir` to top-level
