+++ crates/wasi/src/p2/filesystem.rs
@@ -1,16 +1,31 @@
+use crate::TrappableError;
+use crate::filesystem::File;
 use crate::p2::bindings::filesystem::types;
 use crate::p2::{InputStream, OutputStream, Pollable, StreamError, StreamResult};
-use crate::runtime::{AbortOnDropJoinHandle, spawn_blocking};
-use crate::{DirPerms, FilePerms, OpenMode, TrappableError};
+use crate::runtime::AbortOnDropJoinHandle;
 use anyhow::anyhow;
 use bytes::{Bytes, BytesMut};
 use std::io;
 use std::mem;
-use std::sync::Arc;
 
 pub type FsResult<T> = Result<T, FsError>;
 
 pub type FsError = TrappableError<types::ErrorCode>;
+
+impl From<crate::filesystem::ErrorCode> for types::ErrorCode {
+    fn from(error: crate::filesystem::ErrorCode) -> Self {
+        match error {
+            crate::filesystem::ErrorCode::BadDescriptor => Self::BadDescriptor,
+            crate::filesystem::ErrorCode::NotDirectory => Self::NotDirectory,
+        }
+    }
+}
+
+impl From<crate::filesystem::ErrorCode> for FsError {
+    fn from(error: crate::filesystem::ErrorCode) -> Self {
+        types::ErrorCode::from(error).into()
+    }
+}
 
 impl From<wasmtime::component::ResourceTableError> for FsError {
     fn from(error: wasmtime::component::ResourceTableError) -> Self {
@@ -21,197 +36,6 @@ impl From<wasmtime::component::ResourceTableError> for FsError {
 impl From<io::Error> for FsError {
     fn from(error: io::Error) -> Self {
         types::ErrorCode::from(error).into()
-    }
-}
-
-pub enum Descriptor {
-    File(File),
-    Dir(Dir),
-}
-
-impl Descriptor {
-    pub fn file(&self) -> Result<&File, types::ErrorCode> {
-        match self {
-            Descriptor::File(f) => Ok(f),
-            Descriptor::Dir(_) => Err(types::ErrorCode::BadDescriptor),
-        }
-    }
-
-    pub fn dir(&self) -> Result<&Dir, types::ErrorCode> {
-        match self {
-            Descriptor::Dir(d) => Ok(d),
-            Descriptor::File(_) => Err(types::ErrorCode::NotDirectory),
-        }
-    }
-
-    pub fn is_file(&self) -> bool {
-        match self {
-            Descriptor::File(_) => true,
-            Descriptor::Dir(_) => false,
-        }
-    }
-
-    pub fn is_dir(&self) -> bool {
-        match self {
-            Descriptor::File(_) => false,
-            Descriptor::Dir(_) => true,
-        }
-    }
-}
-
-#[derive(Clone)]
-pub struct File {
-    /// The operating system File this struct is mediating access to.
-    ///
-    /// Wrapped in an Arc because the same underlying file is used for
-    /// implementing the stream types. A copy is also needed for
-    /// [`spawn_blocking`].
-    ///
-    /// [`spawn_blocking`]: Self::spawn_blocking
-    pub file: Arc<cap_std::fs::File>,
-    /// Permissions to enforce on access to the file. These permissions are
-    /// specified by a user of the `crate::p2::WasiCtxBuilder`, and are
-    /// enforced prior to any enforced by the underlying operating system.
-    pub perms: FilePerms,
-    /// The mode the file was opened under: bits for reading, and writing.
-    /// Required to correctly report the DescriptorFlags, because cap-std
-    /// doesn't presently provide a cross-platform equivalent of reading the
-    /// oflags back out using fcntl.
-    pub open_mode: OpenMode,
-
-    allow_blocking_current_thread: bool,
-}
-
-impl File {
-    pub fn new(
-        file: cap_std::fs::File,
-        perms: FilePerms,
-        open_mode: OpenMode,
-        allow_blocking_current_thread: bool,
-    ) -> Self {
-        Self {
-            file: Arc::new(file),
-            perms,
-            open_mode,
-            allow_blocking_current_thread,
-        }
-    }
-
-    /// Execute the blocking `body` function.
-    ///
-    /// Depending on how the WasiCtx was configured, the body may either be:
-    /// - Executed directly on the current thread. In this case the `async`
-    ///   signature of this method is effectively a lie and the returned
-    ///   Future will always be immediately Ready. Or:
-    /// - Spawned on a background thread using [`tokio::task::spawn_blocking`]
-    ///   and immediately awaited.
-    ///
-    /// Intentionally blocking the executor thread might seem unorthodox, but is
-    /// not actually a problem for specific workloads. See:
-    /// - [`crate::p2::WasiCtxBuilder::allow_blocking_current_thread`]
-    /// - [Poor performance of wasmtime file I/O maybe because tokio](https://github.com/bytecodealliance/wasmtime/issues/7973)
-    /// - [Implement opt-in for enabling WASI to block the current thread](https://github.com/bytecodealliance/wasmtime/pull/8190)
-    pub(crate) async fn run_blocking<F, R>(&self, body: F) -> R
-    where
-        F: FnOnce(&cap_std::fs::File) -> R + Send + 'static,
-        R: Send + 'static,
-    {
-        match self.as_blocking_file() {
-            Some(file) => body(file),
-            None => self.spawn_blocking(body).await,
-        }
-    }
-
-    pub(crate) fn spawn_blocking<F, R>(&self, body: F) -> AbortOnDropJoinHandle<R>
-    where
-        F: FnOnce(&cap_std::fs::File) -> R + Send + 'static,
-        R: Send + 'static,
-    {
-        let f = self.file.clone();
-        spawn_blocking(move || body(&f))
-    }
-
-    /// Returns `Some` when the current thread is allowed to block in filesystem
-    /// operations, and otherwise returns `None` to indicate that
-    /// `spawn_blocking` must be used.
-    pub(crate) fn as_blocking_file(&self) -> Option<&cap_std::fs::File> {
-        if self.allow_blocking_current_thread {
-            Some(&self.file)
-        } else {
-            None
-        }
-    }
-}
-
-#[derive(Clone)]
-pub struct Dir {
-    /// The operating system file descriptor this struct is mediating access
-    /// to.
-    ///
-    /// Wrapped in an Arc because a copy is needed for [`spawn_blocking`].
-    ///
-    /// [`spawn_blocking`]: Self::spawn_blocking
-    pub dir: Arc<cap_std::fs::Dir>,
-    /// Permissions to enforce on access to this directory. These permissions
-    /// are specified by a user of the `crate::p2::WasiCtxBuilder`, and
-    /// are enforced prior to any enforced by the underlying operating system.
-    ///
-    /// These permissions are also enforced on any directories opened under
-    /// this directory.
-    pub perms: DirPerms,
-    /// Permissions to enforce on any files opened under this directory.
-    pub file_perms: FilePerms,
-    /// The mode the directory was opened under: bits for reading, and writing.
-    /// Required to correctly report the DescriptorFlags, because cap-std
-    /// doesn't presently provide a cross-platform equivalent of reading the
-    /// oflags back out using fcntl.
-    pub open_mode: OpenMode,
-
-    allow_blocking_current_thread: bool,
-}
-
-impl Dir {
-    pub fn new(
-        dir: cap_std::fs::Dir,
-        perms: DirPerms,
-        file_perms: FilePerms,
-        open_mode: OpenMode,
-        allow_blocking_current_thread: bool,
-    ) -> Self {
-        Dir {
-            dir: Arc::new(dir),
-            perms,
-            file_perms,
-            open_mode,
-            allow_blocking_current_thread,
-        }
-    }
-
-    /// Execute the blocking `body` function.
-    ///
-    /// Depending on how the WasiCtx was configured, the body may either be:
-    /// - Executed directly on the current thread. In this case the `async`
-    ///   signature of this method is effectively a lie and the returned
-    ///   Future will always be immediately Ready. Or:
-    /// - Spawned on a background thread using [`tokio::task::spawn_blocking`]
-    ///   and immediately awaited.
-    ///
-    /// Intentionally blocking the executor thread might seem unorthodox, but is
-    /// not actually a problem for specific workloads. See:
-    /// - [`crate::p2::WasiCtxBuilder::allow_blocking_current_thread`]
-    /// - [Poor performance of wasmtime file I/O maybe because tokio](https://github.com/bytecodealliance/wasmtime/issues/7973)
-    /// - [Implement opt-in for enabling WASI to block the current thread](https://github.com/bytecodealliance/wasmtime/pull/8190)
-    pub(crate) async fn run_blocking<F, R>(&self, body: F) -> R
-    where
-        F: FnOnce(&cap_std::fs::Dir) -> R + Send + 'static,
-        R: Send + 'static,
-    {
-        if self.allow_blocking_current_thread {
-            body(&self.dir)
-        } else {
-            let d = self.dir.clone();
-            spawn_blocking(move || body(&d)).await
-        }
     }
 }
